requirements.txt
asgiref==3.3.1
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
Django==3.1.5
django-environ==0.4.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
ptyprocess==0.7.0
Pygments==2.7.4
PyJWT==2.0.1
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
